requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
cassandra-driver==3.24.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Django==3.1.4
django-cassandra-engine==1.6.1
django-cors-headers==3.5.0
geomet==0.2.1.post1
google==3.0.0
idna==2.10
kwargs==1.0.1
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.26.2
